getting data from files in loop
